pyproject.toml
[project]
name = "pokerbot"
version = "0.1.0"
description = "Бот для отслеживания статистики покерных игр"
authors = [
    {name = "Николай Марков"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=21.11.1,<22.0.0)",
    "pytz (>=2025.1,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
